requirements.txt
google-auth==2.9.0
pandas==1.5.1
numpy==1.23.2
pandas-gbq==0.17.6
scikit-learn==1.1.2
psycopg2==2.9.3
matplotlib==3.5.2
seaborn==0.11.2
openpyxl==3.0.10
ipykernel==6.17.1